Move send_via_kiss to shared root module
