+++ tests/test_kiss.py
@@ -6,7 +6,7 @@ import pytest
 # depends on it at import time.
 pytest.importorskip("psutil")
 
-import telemetry.hub_telemetry as hub_telemetry
+import shared_functions as shared
 
 class DummySocket:
     def __init__(self):
@@ -22,7 +22,7 @@ class TestKissEscaping(unittest.TestCase):
     def _run_send(self, payload):
         dummy = DummySocket()
         with patch('socket.create_connection', return_value=dummy):
-            hub_telemetry.send_via_kiss(payload)
+            shared.send_via_kiss(payload)
         return dummy.sent
 
     def test_no_escape(self):
